feat(websocket)!: replace JSON transport with protobuf (#598)
* feat(websocket)!: replace JSON transport with protobuf

BREAKING CHANGE: EuroScope websocket clients must use the protobuf envelope protocol; JSON clients are no longer supported.

* chore(websocket): skip unused protobuf build targets

---------

+++ backend/internal/services/strip_manual_fpl.go
@@ -130,7 +130,7 @@ func (s *StripService) CreateVFRFPL(ctx context.Context, session int32, req fron
 		Remarks:        req.Remarks,
 		FplType:        req.FPLType,
 		Language:       req.Language,
-		PersonsOnBoard: req.PersonsOnBoard,
+		PersonsOnBoard: int32(req.PersonsOnBoard),
 	})
 
 	return nil
